Authorisation first method

+++ app/Http/Controllers/QuestionsController.php
@@ -64,7 +64,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort('403','Access Denied !!');
+        }
         return view('Questions.edit',compact('question'));
+
+       
     }
 
     /**
@@ -76,6 +81,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403,'Access Denied !!');
+        }
         $question->update($request->all());
 
         return redirect()->route('questions.index')->with('success','Your question has been updated');
@@ -89,6 +97,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+
+        if(\Gate::denies('delete-question',$question)){
+            abort(403,'Access Denied !!');
+        }
         $question->delete();
 
         return redirect()->route('questions.index')->with('success','Your Question has been deleted.');
